Moved ReservationValidation to Validator namespace
