+++ src/Validator/ReservationValidation.php
@@ -1,9 +1,8 @@
 <?php
 
 
-namespace App\Service;
+namespace App\Validator;
 
-use DateTime;
 use Symfony\Component\Validator\Constraints as Assert;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Validator\Validation;
